+++ classes/lib/class-page-title.php
@@ -161,7 +161,7 @@ class Page_Title {
 			<div class="lsx-hero-banner-block wp-block-cover has-background alignfull">
 				<div class="wp-block-cover__inner-container">
 					<div class="lsx-title-block wp-block-group">
-						<div class="wp-block-group__inner-container aligncontent">
+						<div class="wp-block-group__inner-container">
 							<?php $this->lsx_block_title(); ?>
 						</div>
 					</div>