+++ src/extensions/block-button/background-control.js
@@ -12,7 +12,7 @@ const { addFilter } = wp.hooks;
 const { createHigherOrderComponent } = wp.compose;
 const { Fragment } = wp.element;
 const { InspectorControls, PanelColorSettings } = wp.blockEditor;
-const { PanelBody, ToggleControl, TextControl } = wp.components;
+const { PanelBody, ToggleControl, TextControl, SelectControl } = wp.components;
 
 // Enable spacing control on the following blocks
 const enableCustomButton = ["core/button"];
@@ -46,6 +46,18 @@ function addHoverControlAttribute(settings, name) {
 		},
 		borderRadius: {
 			default: 3
+		},
+		buttonFullWidth: {
+			type: "boolean",
+			default: false
+		},
+		buttonDepth: {
+			type: "boolean",
+			default: false
+		},
+		buttonSize: {
+			type: "string",
+			default: ""
 		},
 		buttonModal: {
 			type: "boolean",
@@ -86,11 +98,34 @@ const withHoverControl = createHigherOrderComponent(BlockEdit => {
 			buttonShadowColor,
 			buttonHoverShadowColor,
 			buttonHoverTextColor,
+			buttonFullWidth,
+			buttonDepth,
+			buttonSize,
 			buttonModal,
 			buttonDataTarget,
 			buttonDataToggle
 		} = props.attributes;
 
+		// Button size values
+		const buttonSizeOptions = [
+			{ value: "lsx-button-size-small", label: __("Small") },
+			{ value: "", label: __("Medium") },
+			{ value: "lsx-button-size-large", label: __("Large") }
+		];
+
+		if (buttonFullWidth === true) {
+			var buttonFullWidthClass = "button-full-width-true";
+		} else {
+			var buttonFullWidthClass = "button-full-width-false";
+		}
+		if (buttonDepth === true) {
+			var buttonDepthClass = "button-3d-true";
+		} else {
+			var buttonDepthClass = "button-3d-false";
+		}
+		if (buttonSize) {
+			var buttonSizeClass = buttonSize;
+		}
 		// add has-hover class to block
 		if (buttonHoverColor) {
 			var buttonHoverClass = `has-hover-color-` + buttonHoverColor;
@@ -109,20 +144,101 @@ const withHoverControl = createHigherOrderComponent(BlockEdit => {
 		// add has-shadow class to block
 		props.attributes.className = classnames(
 			props.attributes.className,
+			buttonFullWidthClass,
+			buttonDepthClass,
 			buttonHoverClass,
+			buttonSizeClass,
 			buttonShadowClass,
 			buttonHoverShadowClass,
 			buttonHoverTextClass
 		);
+		//console.log(props.attributes.className);
+
+		var checkWidth = "button-full-width-true";
+		var checkDepth = "button-3d-true";
+		var myPatternWidth = new RegExp("(\\w*" + checkWidth + "\\w*)", "gi");
+		var myPatternDepth = new RegExp("(\\w*" + checkDepth + "\\w*)", "gi");
+
+		var matches = props.attributes.className.match(myPatternWidth);
+		var matchesDepth = props.attributes.className.match(myPatternDepth);
+
 		props.attributes.className = props.attributes.className.split(" ");
 		props.attributes.className = props.attributes.className.filter(onlyUnique);
 		props.attributes.className = props.attributes.className.join(" ");
+
+		if (matches != undefined) {
+			//console.log(matches[0]);
+			if (buttonFullWidth === true) {
+				props.attributes.className = props.attributes.className.replace(
+					"button-full-width-false",
+					""
+				);
+				//console.log(props.attributes.className);
+			} else {
+				props.attributes.className = props.attributes.className.replace(
+					"button-full-width-true",
+					""
+				);
+				//console.log(props.attributes.className);
+			}
+		}
+		if (matchesDepth != undefined) {
+			if (buttonDepth === true) {
+				props.attributes.className = props.attributes.className.replace(
+					"button-3d-false",
+					""
+				);
+			} else {
+				props.attributes.className = props.attributes.className.replace(
+					"button-3d-true",
+					""
+				);
+			}
+		}
+
+		if (buttonSize !== "lsx-button-size-small") {
+			props.attributes.className = props.attributes.className.replace(
+				"lsx-button-size-small",
+				""
+			);
+		}
+		if (buttonSize !== "lsx-button-size-large") {
+			props.attributes.className = props.attributes.className.replace(
+				"lsx-button-size-large",
+				""
+			);
+		}
 
 		return (
 			<Fragment>
 				<BlockEdit {...props} />
 				<InspectorControls>
 					<PanelBody title={__("Additional LSX Settings")} initialOpen={true}>
+						<ToggleControl
+							label={__("Display fullwidth")}
+							checked={buttonFullWidth}
+							onChange={() =>
+								props.setAttributes({ buttonFullWidth: !buttonFullWidth })
+							}
+						/>
+						<ToggleControl
+							label={__("3D Button")}
+							checked={buttonDepth}
+							onChange={() =>
+								props.setAttributes({ buttonDepth: !buttonDepth })
+							}
+						/>
+						<SelectControl
+							label={__("Button Size")}
+							value={buttonSize}
+							options={buttonSizeOptions.map(({ value, label }) => ({
+								value: value,
+								label: label
+							}))}
+							onChange={value => {
+								props.setAttributes({ buttonSize: value });
+							}}
+						/>
 						<PanelColorSettings
 							title={__("Button Hover Color")}
 							initialOpen={true}
@@ -285,6 +401,9 @@ addFilter(
  */
 const addExtraClassesButton = (element, block, attributes) => {
 	let boxShadowStyle = "none";
+	if (attributes.buttonDepth === true) {
+		boxShadowStyle = "2px 2px 0 0 black";
+	}
 	if (attributes.buttonShadowColor) {
 		boxShadowStyle = "2px 2px 0 0 " + attributes.buttonShadowColor;
 	}