+++ src/extensions/block-group/group-control.js
@@ -1,7 +1,7 @@
 // Import CSS
-import './styles/style.scss';
+import "./styles/style.scss";
 
-import assign from 'lodash.assign';
+import assign from "lodash.assign";
 
 const { __ } = wp.i18n;
 
@@ -14,9 +14,7 @@ const { InspectorControls } = wp.blockEditor;
 const { PanelBody, RangeControl, TextControl } = wp.components;
 
 // Enable spacing control on the following blocks
-const enableCustomGroup = [
-	'core/group',
-];
+const enableCustomGroup = ["core/group"];
 
 /**
  * Add the attributes to the group
@@ -25,142 +23,165 @@ const enableCustomGroup = [
  * @param {*} name
  * @returns
  */
-function addGroupControlAttributes ( settings, name ) {
+function addGroupControlAttributes(settings, name) {
 	// Do nothing if it's another block than our defined ones.
-	if ( ! enableCustomGroup.includes( name ) ) {
+	if (!enableCustomGroup.includes(name)) {
 		return settings;
 	}
-	settings.attributes = assign( settings.attributes, {
+	settings.attributes = assign(settings.attributes, {
+		groupMaxWidth: {
+			type: "text"
+		},
+		groupMinHeight: {
+			type: "text"
+		},
 		groupPaddingTop: {
-			type: 'number',
+			type: "text"
 		},
 		groupPaddingBottom: {
-			type: 'number',
+			type: "text"
 		},
 		groupPaddingLeft: {
-			type: 'number',
+			type: "text"
 		},
 		groupPaddingRight: {
-			type: 'number',
+			type: "text"
 		},
 		groupMarginTop: {
-			type: 'number',
+			type: "text"
 		},
 		groupMarginBottom: {
-			type: 'number',
+			type: "text"
 		},
 		groupMarginLeft: {
-			type: 'number',
+			type: "text"
 		},
 		groupMarginRight: {
-			type: 'number',
-		},
-	} );
+			type: "text"
+		}
+	});
 
 	return settings;
 }
 addFilter(
-	'blocks.registerBlockType',
-	'lsx-blocks/extend-group-block/group-control-attribute',
+	"blocks.registerBlockType",
+	"lsx-blocks/extend-group-block/group-control-attribute",
 	addGroupControlAttributes
 );
 
 /**
  * Create the Margin and Padding controls on the inspector controls of block.
  */
-const groupSettingsControl = createHigherOrderComponent( ( BlockEdit ) => {
-	return ( props ) => {
+const groupSettingsControl = createHigherOrderComponent(BlockEdit => {
+	return props => {
 		// Do nothing if it's another block than our defined ones.
-		if ( ! enableCustomGroup.includes( props.name ) ) {
-			return (
-				<BlockEdit { ...props } />
-			);
+		if (!enableCustomGroup.includes(props.name)) {
+			return <BlockEdit {...props} />;
 		}
 
-		const { groupPaddingTop, groupPaddingBottom, groupPaddingLeft, groupPaddingRight, groupMarginTop, groupMarginBottom, groupMarginLeft, groupMarginRight } = props.attributes;
+		const {
+			groupMaxWidth,
+			groupMinHeight,
+			groupPaddingTop,
+			groupPaddingBottom,
+			groupPaddingLeft,
+			groupPaddingRight,
+			groupMarginTop,
+			groupMarginBottom,
+			groupMarginLeft,
+			groupMarginRight
+		} = props.attributes;
 
 		return (
 			<Fragment>
-				<BlockEdit { ...props } />
+				<BlockEdit {...props} />
+
 				<InspectorControls>
-					<PanelBody
-						title={ __( 'Additional LSX Settings' ) }
-						initialOpen={ true }
-					>
-						<RangeControl
-							label={ __( 'Group Padding Top (px)' ) }
-							value={ groupPaddingTop }
-							onChange={ ( value ) => props.setAttributes( { groupPaddingTop: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+					<PanelBody title={__("Additional LSX Settings")} initialOpen={true}>
+						<TextControl
+							label={__("Max width (px)")}
+							type="text"
+							value={groupMaxWidth}
+							onChange={value => props.setAttributes({ groupMaxWidth: value })}
 						/>
-						<RangeControl
-							label={ __( 'Group Padding Bottom (px)' ) }
-							value={ groupPaddingBottom }
-							onChange={ ( value ) => props.setAttributes( { groupPaddingBottom: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Min height (px)")}
+							type="text"
+							value={groupMinHeight}
+							onChange={value => props.setAttributes({ groupMinHeight: value })}
 						/>
-						<RangeControl
-							label={ __( 'Group Padding Left (px)' ) }
-							value={ groupPaddingLeft }
-							onChange={ ( value ) => props.setAttributes( { groupPaddingLeft: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Padding Top (px)")}
+							type="text"
+							value={groupPaddingTop}
+							onChange={value =>
+								props.setAttributes({ groupPaddingTop: value })
+							}
 						/>
-						<RangeControl
-							label={ __( 'Group Padding Right (px)' ) }
-							value={ groupPaddingRight }
-							onChange={ ( value ) => props.setAttributes( { groupPaddingRight: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Padding Bottom (px)")}
+							type="text"
+							value={groupPaddingBottom}
+							onChange={value =>
+								props.setAttributes({ groupPaddingBottom: value })
+							}
 						/>
-						<RangeControl
-							label={ __( 'Group Margin Top (px)' ) }
-							value={ groupMarginTop }
-							onChange={ ( value ) => props.setAttributes( { groupMarginTop: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Padding Left (px)")}
+							type="text"
+							value={groupPaddingLeft}
+							onChange={value =>
+								props.setAttributes({ groupPaddingLeft: value })
+							}
 						/>
-						<RangeControl
-							label={ __( 'Group Margin Bottom (px)' ) }
-							value={ groupMarginBottom }
-							onChange={ ( value ) => props.setAttributes( { groupMarginBottom: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Padding Right (px)")}
+							type="text"
+							value={groupPaddingRight}
+							onChange={value =>
+								props.setAttributes({ groupPaddingRight: value })
+							}
 						/>
-						<RangeControl
-							label={ __( 'Group Margin Left (px)' ) }
-							value={ groupMarginLeft }
-							onChange={ ( value ) => props.setAttributes( { groupMarginLeft: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Margin Top (px)")}
+							type="text"
+							value={groupMarginTop}
+							onChange={value => props.setAttributes({ groupMarginTop: value })}
 						/>
-						<RangeControl
-							label={ __( 'Group Margin Right (px)' ) }
-							value={ groupMarginRight }
-							onChange={ ( value ) => props.setAttributes( { groupMarginRight: value } ) }
-							min={ 0 }
-							max={ 100 }
-							step={ 1 }
+						<TextControl
+							label={__("Group Margin Bottom (px)")}
+							type="text"
+							value={groupMarginBottom}
+							onChange={value =>
+								props.setAttributes({ groupMarginBottom: value })
+							}
+						/>
+						<TextControl
+							label={__("Group Margin Left (px)")}
+							type="text"
+							value={groupMarginLeft}
+							onChange={value =>
+								props.setAttributes({ groupMarginLeft: value })
+							}
+						/>
+						<TextControl
+							label={__("Group Margin Right (px)")}
+							type="text"
+							value={groupMarginRight}
+							onChange={value =>
+								props.setAttributes({ groupMarginRight: value })
+							}
 						/>
 					</PanelBody>
 				</InspectorControls>
 			</Fragment>
 		);
 	};
-}, 'groupSettingsControl' );
+}, "groupSettingsControl");
 
 addFilter(
-	'editor.BlockEdit',
-	'lsx-blocks/extend-group-block/group-settings-control',
+	"editor.BlockEdit",
+	"lsx-blocks/extend-group-block/group-settings-control",
 	groupSettingsControl
 );
 
@@ -173,35 +194,127 @@ addFilter(
  *
  * @returns {object} Modified props of save element.
  */
-const addExtraPropsGroup = ( saveElementProps, blockType, attributes ) => {
-
+const addExtraPropsGroup = (saveElementProps, blockType, attributes) => {
 	// Do nothing if it's another block than our defined ones.
-	if ( ! enableCustomGroup.includes( blockType.name ) ) {
+	if (!enableCustomGroup.includes(blockType.name)) {
 		return saveElementProps;
 	}
+	// console.log(attributes.groupPaddingTop);
+	// console.log(attributes.groupPaddingBottom);
+	if (blockType.name === "core/group") {
+		let paddingTop = attributes.groupPaddingTop + "px";
+		let paddingBottom = attributes.groupPaddingBottom + "px";
+		let paddingLeft = attributes.groupPaddingLeft + "px";
+		let paddingRight = attributes.groupPaddingRight + "px";
+		let marginTop = attributes.groupMarginTop + "px";
+		let marginBottom = attributes.groupMarginBottom + "px";
+		let marginLeft = attributes.groupMarginLeft + "px";
+		let marginRight = attributes.groupMarginRight + "px";
 
-	if ( blockType.name === 'core/group' ) {
+		let str = "";
+		let style = "";
+		var obj2 = "";
 
-		assign( saveElementProps,
-			{ style: {
-				'paddingTop': attributes.groupPaddingTop,
-				'paddingBottom': attributes.groupPaddingBottom,
-				'paddingLeft': attributes.groupPaddingLeft,
-				'paddingRight': attributes.groupPaddingRight,
-				'marginTop': attributes.groupMarginTop,
-				'marginBottom': attributes.groupMarginBottom,
-				'marginLeft': attributes.groupMarginLeft,
-				'marginRight': attributes.groupMarginRight,
-			}}
-		);
+		if (undefined !== attributes.groupPaddingTop) {
+			var obj2 = { ...obj2, paddingTop: paddingTop };
+		}
+		if (undefined !== attributes.groupPaddingBottom) {
+			var obj2 = { ...obj2, paddingBottom: paddingBottom };
+		}
+		if (undefined !== attributes.groupPaddingLeft) {
+			var obj2 = { ...obj2, paddingLeft: paddingLeft };
+		}
+		if (undefined !== attributes.groupPaddingRight) {
+			var obj2 = { ...obj2, paddingRight: paddingRight };
+		}
+		if (undefined !== attributes.groupMarginTop) {
+			var obj2 = { ...obj2, marginTop: marginTop };
+		}
+		if (undefined !== attributes.groupMarginBottom) {
+			var obj2 = { ...obj2, marginBottom: marginBottom };
+		}
+		if (undefined !== attributes.groupMarginLeft) {
+			var obj2 = { ...obj2, marginLeft: marginLeft };
+		}
+		if (undefined !== attributes.groupMarginRight) {
+			var obj2 = { ...obj2, marginRight: marginRight };
+		}
+
+		str = saveElementProps.style;
+
+		style = { ...str, ...obj2 };
+
+		// console.log(saveElementProps);
+		// console.log(attributes);
+
+		assign(saveElementProps, { style });
 	}
 
 	return saveElementProps;
-
 };
 
 addFilter(
-	'blocks.getSaveContent.extraProps',
-	'lsx-blocks/extend-group-block/add-extra-props-group',
+	"blocks.getSaveContent.extraProps",
+	"lsx-blocks/extend-group-block/add-extra-props-group",
 	addExtraPropsGroup
+);
+
+/**
+ * Add box shadow styles to button.
+ *
+ * @param {object} saveElementProps Props of save element.
+ * @param {Object} blockType Block type information.
+ * @param {Object} attributes Attributes of block.
+ *
+ * @returns {object} Modified props of save element.
+ */
+const addChildrenStyles = (element, block, attributes) => {
+	let maxWidth = "";
+	let minHeight = "";
+	if (attributes.groupMaxWidth) {
+		maxWidth = attributes.groupMaxWidth + "px";
+	}
+
+	if (attributes.groupMinHeight) {
+		minHeight = attributes.groupMinHeight + "px";
+	}
+
+	if (block.name === "core/group" && undefined !== attributes.groupMaxWidth) {
+		//console.log(attributes.groupMaxWidth);
+		let style = "";
+		var obj2 = { marginLeft: "auto", marginRight: "auto" };
+		var savedElement = undefined;
+
+		if (undefined !== attributes.groupMaxWidth) {
+			var obj2 = { ...obj2, maxWidth: maxWidth };
+		}
+		if (undefined !== attributes.groupMinHeight) {
+			var obj2 = { ...obj2, minHeight: minHeight };
+		}
+		//console.log(element.props.children.props.className);
+		if (undefined !== element.props.children) {
+			savedElement = element.props.children;
+		}
+
+		//console.log(savedElement);
+		if (undefined !== savedElement) {
+			style = { ...obj2 };
+
+			return wp.element.cloneElement(
+				element,
+				{},
+				wp.element.cloneElement(savedElement, {
+					style
+				})
+			);
+		}
+	}
+
+	return element;
+};
+
+addFilter(
+	"blocks.getSaveElement",
+	"lsx-blocks/extend-group-block-add-child-styles",
+	addChildrenStyles
 );